Bring tool runtime tests back under the architecture budget
The governance lane enforces the structural-size budget for crates/app/src/tools/mod.rs.
After the recovery-hardening follow-up landed, that file crossed the 15k-line
limit and strict architecture checks started failing. The lowest-risk repair is
to move a large contiguous shell/tool-search test cluster into a child test file
inside the same module, keeping helper visibility and behavior unchanged while
shrinking the production file back under budget.

The monthly architecture drift report was regenerated in the same change so the
tracked hotspot snapshot matches the new file layout and the strict governance
checks stop flagging stale metadata.

Constraint: Governance requires crates/app/src/tools/mod.rs to stay within the tracked structural-size budget
Constraint: Keep the runtime behavior unchanged while repairing the file-size regression
Rejected: Refactor runtime logic into new production modules just to satisfy the size budget | larger functional risk than moving test-only code
Rejected: Ignore the stale drift report until after merge | governance blocks the PR immediately when the tracked report is out of date
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: When tools/mod.rs approaches the budget again, split additional test-only coverage before moving runtime logic unless the runtime seams themselves truly need extraction
Tested: cargo fmt --all -- --check
Tested: LOONGCLAW_ARCH_STRICT=true ./scripts/check_architecture_boundaries.sh
Tested: bash scripts/check_architecture_drift_freshness.sh docs/releases/architecture-drift-2026-04.md
Tested: cargo test -p loongclaw-app --all-features keywords_array_payloads
Not-tested: the full workspace suites were not rerun after this test-only split because the change is architecture/report-only and earlier app/runtime checks had already passed
